# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: KatalogitAI API
  version: 1.0.0
  description: >
    Backend for KatalogitAI — AI product photoshoots for fashion vendors.
    All endpoints (except webhooks) require `Authorization: Bearer <Firebase ID token>`.
    The credit ledger lives in Cloud SQL; catalogue lives in Firestore.
servers:
  - url: /api/v1
    description: Same-origin via Firebase Hosting rewrite to Cloud Run
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
        data: { type: object }
      required: [ok]
...
    Store:
      type: object
      properties:
        id: { type: string }
        ownerUid: { type: string }
        name: { type: string }
        city: { type: string }
        categories: { type: array, items: { type: string } }
        scale: { type: string, enum: [starter, small, growing, large] }
        aiFeatures: { type: array, items: { type: string } }
        plan: { type: string, enum: [free, pro, business] }
        status: { type: string, enum: [active, suspended] }
...
    UploadUrlRequest:
      type: object
      required: [fileName, contentType, purpose]
      properties:
        fileName: { type: string }
        contentType: { type: string, enum: [image/jpeg, image/png, image/webp] }
        purpose: { type: string, enum: [flat_lay, ai_result] }
    UploadUrlResponse:
      type: object
      properties:
        uploadUrl: { type: string, description: V4 signed PUT URL, 15-min TTL }
        objectPath: { type: string }
...
  responses:
    Ok:
      description: Success
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  data: { type: object }
...
security:
  - bearerAuth: []
paths:
  /stores:
    post:
      summary: Create a store (onboarding completion)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Store'
      responses:
        '201': { $ref: '#/components/responses/Ok' }
        '409': { description: Store already exists for this user }
  /stores/{storeId}:
    get:
      summary: Read store profile + plan
      parameters: [{ $in: path, name: storeId, required: true, schema: { type: string } }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    patch:
      summary: Update store profile
      parameters: [{ $in: path, name: storeId, required: true, schema: { type: string } }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /uploads/urls:
    post:
      summary: Request a signed upload URL (direct-to-GCS)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UploadUrlRequest' }
      responses:
        '200':
          description: Signed upload URL
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/UploadUrlResponse' }
...